Add pylons/flask mixin
