+++ ckanext/report/plugin/__init__.py
@@ -1,4 +1,5 @@
 import ckan.plugins as p
+from ckan.plugins import toolkit
 from ckanext.report.interfaces import IReport
 
 import ckanext.report.logic.action.get as action_get
@@ -6,30 +7,24 @@ import ckanext.report.logic.action.update as action_update
 import ckanext.report.logic.auth.get as auth_get
 import ckanext.report.logic.auth.update as auth_update
 
-class ReportPlugin(p.SingletonPlugin):
+try:
+    toolkit.requires_ckan_version("2.9")
+except CkanVersionException:
+    from ckanext.report.plugin.pylons_plugin import MixinPlugin
+else:
+    from ckanext.report.plugin.flask_plugin import MixinPlugin
+
+class ReportPlugin(MixinPlugin, p.SingletonPlugin):
     p.implements(p.IRoutes, inherit=True)
     p.implements(p.IConfigurer)
     p.implements(p.ITemplateHelpers)
     p.implements(p.IActions, inherit=True)
     p.implements(p.IAuthFunctions, inherit=True)
 
-    # IRoutes
-
-    def before_map(self, map):
-        report_ctlr = 'ckanext.report.controllers:ReportController'
-        map.connect('reports', '/report', controller=report_ctlr,
-                    action='index')
-        map.redirect('/reports', '/report')
-        map.connect('report', '/report/:report_name', controller=report_ctlr,
-                    action='view')
-        map.connect('report-org', '/report/:report_name/:organization',
-                    controller=report_ctlr, action='view')
-        return map
-
     # IConfigurer
 
     def update_config(self, config):
-        p.toolkit.add_template_directory(config, 'templates')
+        p.toolkit.add_template_directory(config, '../templates')
 
     # ITemplateHelpers
