+++ ckanext/report/plugin/pylons_plugin.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+
+import ckan.plugins as p
+
+class MixinPlugin(p.SingletonPlugin):
+    p.implements(p.IRoutes, inherit=True)
+
+    # IRoutes
+
+    def before_map(self, map):
+        report_ctlr = 'ckanext.report.controllers:ReportController'
+        map.connect('reports', '/report', controller=report_ctlr,
+                    action='index')
+        map.redirect('/reports', '/report')
+        map.connect('report', '/report/:report_name', controller=report_ctlr,
+                    action='view')
+        map.connect('report-org', '/report/:report_name/:organization',
+                    controller=report_ctlr, action='view')
+        return map
